+++ tests/test_data_types.py
@@ -41,6 +41,11 @@ class TestText(unittest.TestCase):
         self.assertEqual(self.type.test('a'), True)
         self.assertEqual(self.type.test('A\nB'), True)
         self.assertEqual(self.type.test(u'👍'), True)
+        self.assertEqual(self.type.test('05_leslie3d_base'), True)
+        self.assertEqual(self.type.test('2016-12-29'), True)
+        self.assertEqual(self.type.test('2016-12-29T11:43:30Z'), True)
+        self.assertEqual(self.type.test('2016-12-29T11:43:30+06:00'), True)
+        self.assertEqual(self.type.test('2016-12-29T11:43:30-06:00'), True)
 
     def test_cast(self):
         values = ('a', 1, None, Decimal('2.7'), 'n/a', u'👍', ' foo', 'foo ')
@@ -86,6 +91,11 @@ class TestBoolean(unittest.TestCase):
         self.assertEqual(self.type.test('a'), False)
         self.assertEqual(self.type.test('A\nB'), False)
         self.assertEqual(self.type.test(u'👍'), False)
+        self.assertEqual(self.type.test('05_leslie3d_base'), False)
+        self.assertEqual(self.type.test('2016-12-29'), False)
+        self.assertEqual(self.type.test('2016-12-29T11:43:30Z'), False)
+        self.assertEqual(self.type.test('2016-12-29T11:43:30+06:00'), False)
+        self.assertEqual(self.type.test('2016-12-29T11:43:30-06:00'), False)
 
     def test_cast(self):
         values = (True, 'yes', None, False, 'no', 'n/a', '1', 0)
@@ -130,6 +140,11 @@ class TestNumber(unittest.TestCase):
         self.assertEqual(self.type.test('a'), False)
         self.assertEqual(self.type.test('A\nB'), False)
         self.assertEqual(self.type.test(u'👍'), False)
+        self.assertEqual(self.type.test('05_leslie3d_base'), False)
+        self.assertEqual(self.type.test('2016-12-29'), False)
+        self.assertEqual(self.type.test('2016-12-29T11:43:30Z'), False)
+        self.assertEqual(self.type.test('2016-12-29T11:43:30+06:00'), False)
+        self.assertEqual(self.type.test('2016-12-29T11:43:30-06:00'), False)
 
     def test_cast(self):
         values = (2, 1, None, Decimal('2.7'), 'n/a', '2.7', '200,000,000')
@@ -187,6 +202,11 @@ class TestDate(unittest.TestCase):
         self.assertEqual(self.type.test('a'), False)
         self.assertEqual(self.type.test('A\nB'), False)
         self.assertEqual(self.type.test(u'👍'), False)
+        self.assertEqual(self.type.test('05_leslie3d_base'), False)
+        self.assertEqual(self.type.test('2016-12-29'), True)
+        self.assertEqual(self.type.test('2016-12-29T11:43:30Z'), False)
+        self.assertEqual(self.type.test('2016-12-29T11:43:30+06:00'), False)
+        self.assertEqual(self.type.test('2016-12-29T11:43:30-06:00'), False)
 
     def test_test_format(self):
         date_type = Date(date_format='%m-%d-%Y')
@@ -260,6 +280,11 @@ class TestDateTime(unittest.TestCase):
         self.assertEqual(self.type.test('a'), False)
         self.assertEqual(self.type.test('A\nB'), False)
         self.assertEqual(self.type.test(u'👍'), False)
+        self.assertEqual(self.type.test('05_leslie3d_base'), False)
+        self.assertEqual(self.type.test('2016-12-29'), True)
+        self.assertEqual(self.type.test('2016-12-29T11:43:30Z'), True)
+        self.assertEqual(self.type.test('2016-12-29T11:43:30+06:00'), True)
+        self.assertEqual(self.type.test('2016-12-29T11:43:30-06:00'), True)
 
     def test_test_format(self):
         datetime_type = DateTime(datetime_format='%m-%d-%Y %I:%M %p')
@@ -350,6 +375,11 @@ class TestTimeDelta(unittest.TestCase):
         self.assertEqual(self.type.test('a'), False)
         self.assertEqual(self.type.test('A\nB'), False)
         self.assertEqual(self.type.test(u'👍'), False)
+        self.assertEqual(self.type.test('05_leslie3d_base'), False)
+        self.assertEqual(self.type.test('2016-12-29'), False)
+        self.assertEqual(self.type.test('2016-12-29T11:43:30Z'), False)
+        self.assertEqual(self.type.test('2016-12-29T11:43:30+06:00'), False)
+        self.assertEqual(self.type.test('2016-12-29T11:43:30-06:00'), False)
 
     def test_cast_parser(self):
         values = ('4:10', '1.2m', '172 hours', '5 weeks, 2 days', 'n/a')